+++ packages/ricos-content/web/src/converters/proseMirror/utils.ts
@@ -1,5 +1,8 @@
 import { capitalize } from 'lodash';
+import { Node_Type } from 'ricos-schema';
 import { DECORATION_TYPES, NODE_TYPES } from './consts';
+
+const TYPES = [...NODE_TYPES, ...DECORATION_TYPES];
 
 export const isNode = object =>
   NODE_TYPES.includes(object?.type?.toUpperCase()) && ('nodes' in object || 'content' in object);
@@ -13,3 +16,8 @@ export const toDataFieldName = (type: string) =>
     .map((str, index) => (index === 0 ? str : capitalize(str)))
     .join('')
     .concat('Data');
+
+export const DATA_FIELDS_MAP = {
+  ...TYPES.reduce((map, type) => ({ ...map, [type]: toDataFieldName(type) }), {}),
+  [Node_Type.CODEBLOCK]: 'codeData',
+};